Move the guard to where it guards, and stop the compileOnly test lying
Review findings on the previous commit.

The "logging still works without the Compose runtime" test was vacuous where it
lived. plugin-loader gets the Compose runtime transitively (plugin-api-core
api-exposes it), so the exact condition the test claimed to verify — "absent here
exactly as in a packaged host" — was false and it asserted nothing.

Moved the whole test into plugin-logging's own desktopTest, where compileOnly deps
really are off the test runtime classpath. Confirmed by dumping
desktopTestRuntimeClasspath: kotlin-stdlib, coroutines, slf4j, kotlin-test and no
Compose. That also closes the second finding — the guard now lives in the module it
guards, so it is on the publish path (publishAndReleaseToMavenCentral's task graph
did not include plugin-loader's tests, so a future 1.0.6 could have been published
without $stable and nothing on that path would have objected).

Extended typesNeedingStable from two classes to all five public non-enum types in
the package — LogEntry, BossLoggerConfig and LogSanitizer get $stable too, and a
plugin holding any of them as a property would have broken identically. The KDoc
implied the two-entry list was exhaustive. Also documented that only presence is
asserted and not the $stable value: the bitmask is computed independently by each
repo's Compose version, and a mismatch cannot break linking, only recomposition
skipping.

Moved compileOnly(compose runtime) from commonMain — which has no Kotlin sources at
all — into desktopMain, next to the classes the Compose compiler actually processes.

Added a targeted diagnostic to BinaryCompatibilityValidator: a missing field named
$stable now says so in the error, because that field is Compose-synthetic and its
absence almost always means two copies of the owning class exist, one compiled with
Compose and one without. "$stable: field not found" being the only clue is what made
this take hours. Deliberately a message change, NOT a soft-fail — the plugin's
getstatic sits in its own <clinit>, so a genuinely missing field is a hard
NoSuchFieldError at first touch; rejecting up front with a reason beats crashing
later without one.

Not in this PR, filed as a follow-up: reading the logging classes out of the
unfiltered api jar that plugin-api-core already downloads to build/api-contract/ and
diffing public members against this module's copy. That is the guard that catches the
NEXT divergence rather than this one, and it deserves its own change.
